mp-filament-sensor: Extend the filament tube
The gap between the filament tube and the start of the bowden
extruder is a problem when the filament has started to run out.

Because of buffering, the print continues after the sensor has been
triggered.  While continuing, it can cause the filament to smash
into the side of the sensor module.  When smashing into the side,
we are distorting the filament which seems like a bad idea when
it may end up entering the extruder before the printing really
stops.