mp-azteeg-side-panel:
Custom left side panel for the Azteeg X5 GT mounted in a monoprice
select mini.

I also generated a right side plate with a location to mount a fan
so I could relocate the rear fan to allow me to use the existing
fan hole to reroute the heater wires.