mp-e3d-part-fan: Adjust air exit holes to try to balance airflow better
Make the two holes closest to the fan smaller and leave more of a wall
directly in the middle of the air path to slow down the direct blowing.

Make the far side into a single larger hole to try to get the maximal
amount of air exiting on the side.