palette-tube-to-ptfe: Adaptor tweaks and add an end gripper
I really, really hate trying to get the ptfe tube into the idler
arm of the titan extruder and even if I do get it in sometimes it
retracts back out.  Worse it then sometimes pulls itself back
in.  This makes it impossible for the palette to compensate
reasonably for these changes.

The end gripper gives me a consistent point for the tube to
stop at the extruder.  Using this does require that you use my
version of gcode2msf and that you enable pings_ignore_retraction.

The adaptor is tweaked for tube dimensions and to include a filament
guide in the connection point to help guide it through cleanly.